+++ src/App.tsx
@@ -8,7 +8,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { CartProvider, useCart } from './context/CartContext';
 import CartModal from './components/CartModal';
-import Login from './pages/Login';
+import Login from './pages/LoginPage';
 import MyPage from './pages/MyPage';
 
 const queryClient = new QueryClient();
